requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
yfinance==0.1.63
requests==2.26.0
questdb==1.0.0
beautifulsoup4==4.9.3
scikit-learn==0.24.2
textblob==0.15.3
python-dotenv==0.19.0
httpx==0.19.0
aiofiles==0.8.0
python-multipart==0.0.5
starlette==0.14.2
psycopg2-binary==2.9.3
jinja2==3.0.1